simple_fast_api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
aiohttp
beautifulsoup4
matplotlib
lxml
python-telegram-bot>=20.0
aiosqlite>=0.19
apscheduler>=3.10
diskcache>=5.6.3
yfinance>=0.2
